mcp_client/requirements.txt
mcp
openai
python-dotenv

# Python 3.11.4